Move get_backup_sources to CacheAPI
